requirements.txt
Django==1.4.1
django-tastypie==0.9.12-alpha